+++ CHANGELOG.md
@@ -6,8 +6,8 @@
   * Bumping native-checkout version to 0.8.1 
   * Adding in Native checkout support for one time password
 * BraintreeCore
-  * Add `BraintreeClient#launchesBrowserSwitchAsNewTask()` boolean flag to opt in to using the SDKs default deep link handler
-  * Create `BraintreeDeepLinkActivity` to capture deep link results when default deep link behavior is enabled
+  * Add `BraintreeClient#launchesBrowserSwitchAsNewTask()` boolean flag to allow the SDK to capture deep link results on behalf of the host application
+  * Create `BraintreeDeepLinkActivity` to capture deep link results on behalf of the host application
 
 ## 4.15.0